Archived/mcdonalds/pyproject.toml
[project]
name = "mcdonalds"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pytest-playwright>=0.7.1",
]
